+++ podaac/controllers/default_controller.py
@@ -1,7 +1,7 @@
 import connexion
 import six
 
-from swagger_server import util
+from podaac import util
 
 
 import logging
