+++ podaac/test/test_default_controller.py
@@ -5,7 +5,7 @@ from __future__ import absolute_import
 from flask import json
 from six import BytesIO
 
-from swagger_server.test import BaseTestCase
+from podaac.test import BaseTestCase
 
 
 class TestDefaultController(BaseTestCase):
