+++ setup.py
@@ -3,7 +3,7 @@
 import sys
 from setuptools import setup, find_packages
 
-NAME = "swagger_server"
+NAME = "podaac"
 VERSION = "1.0.0"
 # To install the library, run the following
 #
@@ -29,7 +29,7 @@ setup(
     package_data={'': ['swagger/swagger.yaml']},
     include_package_data=True,
     entry_points={
-        'console_scripts': ['swagger_server=swagger_server.__main__:main']},
+        'console_scripts': ['podaac=podaac.__main__:main']},
     long_description="""\
     Get time series data from SWOT observations for reaches, nodes, and/or lakes
     """
